+header #links
Made all page headings have links to their ids. Also, when the user hovers over a heading, a # will appear before the element as it turns orange with an underline to further drive in the idea that these are clickable sub links.

Additionally non-index SASS files were prefixed with an underscore to prevent them from being written to the output folder (as their contents were already compiled into `index.css`).